requirements.txt
aiolimiter==1.1.0
alembic==1.13.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.10.4
blinker==1.8.2
boto3==1.35.6
botocore==1.35.6
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.3
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-ai-generativelanguage==0.6.10
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.30.0
google-generativeai==0.8.3
Pillow
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx
idna==3.6
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
limits==3.13.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.18.0
pyparsing==3.2.3
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot>=21.0
pytz==2023.3.post1
qrcode==7.4.2
requests==2.31.0
rich==13.8.1
rsa==4.9.1
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.32
telegram==0.0.1
tenacity==9.1.2
tornado==6.3.3
tqdm==4.67.1
types-Flask-Migrate==4.0.0.20240311
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.2
uritemplate==4.2.0
urllib3==2.1.0
websockets==15.0.1
Werkzeug==3.0.3
wrapt==1.16.0
WTForms==3.1.2
